Move wasm queries to modules/wasm folder
