+++ packages/cosmwasm-stargate/src/modules/index.ts
@@ -1,0 +1,1 @@
+export { JsonObject, setupWasmExtension, WasmExtension } from "./wasm/queries";
