--- packages/cosmwasm-stargate/src/queries/index.ts
@@ -1,1 +1,0 @@
-export { JsonObject, setupWasmExtension, WasmExtension } from "./wasm";
